moved mockStorageHAL.hpp to nativeMocksAndHelpers
